feat(backend): 鉴权/支付/订单/智能体等核心服务增强与 P3 测试补齐
- auth/security/deps: 鉴权与依赖注入加固
- payment/payment_providers/order_service: 支付与订单流转（escrow/pickup）完善
- agent/ai_features/recommendation/vision: 智能体与推荐能力增强
- coupon/points/replenishment/support/auto_review: 营销与履约服务优化
- state_machine 与模型扩展
- 补齐 agent/escrow/payments/pickup P3 测试

+++ backend/app/models/presale.py
@@ -3,7 +3,7 @@ import enum
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import Column, DateTime, Enum as SAEnum, Float, ForeignKey, Integer, Numeric, String
+from sqlalchemy import Column, DateTime, Enum as SAEnum, ForeignKey, Integer, Numeric, String
 
 from app.db.base import Base
 
@@ -27,7 +27,7 @@ class Presale(Base):
     title = Column(String(100), nullable=False)
     presale_price = Column(Numeric(12, 2), nullable=False)
     deposit = Column(Numeric(12, 2), nullable=False)
-    inflate_rate = Column(Float, default=1.5, nullable=False)
+    inflate_rate = Column(Numeric(4, 2), default=1.5, nullable=False)
     end_at = Column(DateTime(timezone=True), nullable=True)
     is_active = Column(Integer, default=1, nullable=False)
     created_at = Column(DateTime(timezone=True), default=_now)
